refactor(KeyboardShortcuts): rename from "Keyboard"
"Keyboard" is a bit too generic/ambiguous
