+++ specifyweb/frontend/js_src/lib/components/KeyboardShortcuts/Shortcuts.tsx
