+++ specifyweb/frontend/js_src/lib/components/KeyboardShortcuts/config.ts
